re-organise serialised objects, add resurrect events, bump version number
