+++ .github/workflows/release.yml
@@ -9,7 +9,7 @@ jobs:
   build:
     strategy:
       matrix:
-        platform: [macos-latest, windows-latest, ubuntu-latest]
+        platform: [macos-latest, windows-latest]
     runs-on: ${{ matrix.platform }}
     defaults:
       run:
@@ -19,19 +19,19 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
       
-      - name: Install Linux dependencies
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt update
-          sudo apt install -y \
-            pkg-config \
-            libglib2.0-dev \
-            libgtk-3-dev \
-            libayatana-appindicator3-dev \
-            librsvg2-dev \
-            libsoup-3.0-dev \
-            libjavascriptcoregtk-4.1-dev \
-            libwebkit2gtk-4.1-dev
+      # - name: Install Linux dependencies
+      #   if: runner.os == 'Linux'
+      #   run: |
+      #     sudo apt update
+      #     sudo apt install -y \
+      #       pkg-config \
+      #       libglib2.0-dev \
+      #       libgtk-3-dev \
+      #       libayatana-appindicator3-dev \
+      #       librsvg2-dev \
+      #       libsoup-3.0-dev \
+      #       libjavascriptcoregtk-4.1-dev \
+      #       libwebkit2gtk-4.1-dev
             
       - name: Set up Rust
         uses: dtolnay/rust-toolchain@stable
