+++ cli/Cargo.toml
@@ -10,6 +10,4 @@ ftail = "0.3.0"
 log = "0.4.27"
 parser = { path = "../parser" }
 rayon = "1.11.0"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
 zip = { version = "4.2.0", default-features = false, features = ["deflate"] }
