+++ desktop/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "esologtool-ui"
-version = "0.3.8"
+version = "0.3.9"
 edition = "2021"
 
 [dependencies]
@@ -16,7 +16,6 @@ serde-wasm-bindgen = "0.6"
 console_error_panic_hook = "0.1.7"
 gloo-utils = "0.2.0"
 parser = { path = "../parser" }
-cli = { path = "../cli" }
 yew_icons = {version = "0.8", features = ["LucideArrowLeftCircle", "LucideUpload", "BootstrapFileEarmarkBreak", "BootstrapGithub", "BootstrapFileEarmarkBarGraph", "BootstrapFiles", "BootstrapFolder", "BootstrapFileEarmarkArrowUp", "BootstrapFolderSymlink", "BootstrapPersonCircle", "LucideLogOut", "BootstrapXLg", "BootstrapDiscord", "BootstrapTrash3", "BootstrapFileEarmarkPlay"]}
 stylist = {version = "0.13", features = ["yew_integration"]}
 tauri-sys = { git = "https://github.com/JonasKruckenberg/tauri-sys", branch = "v2", features = ["all"] }
