+++ desktop/src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "ESO Log Tool",
-  "version": "0.3.8",
+  "version": "0.3.9",
   "identifier": "com.esologtool.app",
   "build": {
     "beforeDevCommand": "trunk serve",
