+++ desktop/src/app.rs
@@ -1,6 +1,5 @@
 use std::rc::Rc;
 
-use cli::esologs_format::{LoginResponse, UploadSettings};
 use yew::prelude::*;
 use yew_router::prelude::*;
 use crate::routes::Route;
@@ -12,6 +11,7 @@ use crate::ui::modify::ModifyScreen;
 use crate::ui::split::SplitCombineScreen;
 use crate::ui::terms::TermsComponent;
 use crate::ui::upload::UploadScreen;
+use crate::{LoginResponse, UploadSettings};
 
 pub type LoginContext = UseStateHandle<Option<Rc<LoginResponse>>>;
 pub type ESOLogUploadSettings = UseStateHandle<Option<Rc<UploadSettings>>>;
