+++ desktop/src/main.rs
@@ -1,8 +1,4 @@
-mod app;
-mod routes;
-mod ui;
-
-use app::App;
+use esologtool_ui::app::App;
 
 fn main() {
     console_error_panic_hook::set_once();
