+++ desktop/src/ui/login.rs
@@ -1,12 +1,11 @@
 use std::rc::Rc;
-use cli::esologs_format::LoginResponse;
 use stylist::{css, Style};
 use tauri_sys::core::{invoke, invoke_result};
 use web_sys::HtmlInputElement;
 use yew::prelude::*;
 use yew_router::hooks::use_navigator;
 use yew_icons::{Icon, IconId};
-use crate::{app::LoginContext, routes::Route, ui::{icon_button::{BackArrow, IconButton}, style::*}};
+use crate::{app::LoginContext, routes::Route, ui::{icon_button::{BackArrow, IconButton}, style::*}, LoginResponse};
 
 #[function_component(LoginBox)]
 pub fn login_component() -> Html {
